LayoutTests/css3/flexbox/multiline-line-pack-horizontal-column.html

<!DOCTYPE html>
<html>
<style>
.flexbox {
    position: relative;
    display: -webkit-flex;
    background-color: grey;
    width: 600px;
    height: 20px;
}
.title {
    margin-top: 1em;
}
.ltr {
    direction: ltr;
}
.rtl {
    direction: rtl;
}
.horizontal-tb {
    -webkit-writing-mode: horizontal-tb;
}
.horizontal-bt {
    -webkit-writing-mode: horizontal-bt;
}
.column {
    -webkit-flex-flow: column;
}
.column-reverse {
    -webkit-flex-flow: column-reverse;
}
.wrap {
    -webkit-flex-wrap: wrap;
}
.wrap-reverse {
    -webkit-flex-wrap: wrap-reverse;
}
.line-pack-start {
    -webkit-flex-line-pack: start;
}
.line-pack-end {
    -webkit-flex-line-pack: end;
}
.line-pack-center {
    -webkit-flex-line-pack: center;
}
.line-pack-justify {
    -webkit-flex-line-pack: justify;
}
.line-pack-distribute {
    -webkit-flex-line-pack: distribute;
}
.line-pack-stretch {
    -webkit-flex-line-pack: stretch;
}
.flexbox > :nth-child(1) {
    background-color: blue;
}
.flexbox > :nth-child(2) {
    background-color: green;
}
</style>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<script src="resources/flexbox.js"></script>
<body onload="checkFlexBoxen()">

<script>
var expectations = {
    'horizontal-tb': {
        'column': {
            'ltr': {
                'wrap': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
                'wrap-reverse': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
            },
            'rtl': {
                'wrap': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
                'wrap-reverse': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
            },
        },
        // Same as column.
        'column-reverse': {
            'ltr': {
                'wrap': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
                'wrap-reverse': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
            },
            'rtl': {
                'wrap': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
                'wrap-reverse': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
            },
        },
    },
    'horizontal-bt': {
        // Same as horizontal-tb.
        'column': {
            'ltr': {
                'wrap': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
                'wrap-reverse': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
            },
            'rtl': {
                'wrap': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
                'wrap-reverse': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
            },
        },
        'column-reverse': {
            'ltr': {
                'wrap': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
                'wrap-reverse': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
            },
            'rtl': {
                'wrap': {
                    'start': [500, 400],
                    'end': [100, 0],
                    'center': [300, 200],
                    'justify': [500, 0],
                    'distribute': [400, 100],
                    'stretch': [300, 0],
                },
                'wrap-reverse': {
                    'start': [0, 100],
                    'end': [400, 500],
                    'center': [200, 300],
                    'justify': [0, 500],
                    'distribute': [100, 400],
                    'stretch': [0, 300],
                },
            },
        },
    },
};

function mainAxisDirection(writingMode, flexDirection)
{
    if ((writingMode.indexOf('horizontal') != -1 && flexDirection.indexOf('row') != -1)
        || (writingMode.indexOf('vertical') != -1 && flexDirection.indexOf('column') != -1))
        return 'width';
    return 'height';
}

function addChild(flexbox, preferredSize, linePack, expected_x_offset)
{
    var child = document.createElement('div');
    child.setAttribute('style', '-webkit-flex: 1 ' + preferredSize + 'px;'
        + 'min-width: 100px');

    child.setAttribute("data-expected-width", linePack == "stretch" ? "300" : "100");
    child.setAttribute("data-expected-height", "20");
    child.setAttribute("data-offset-y", "0");
    child.setAttribute("data-offset-x", expected_x_offset);

    flexbox.appendChild(child);
}

var writingModes = ['horizontal-tb', 'horizontal-bt'];
var flexDirections = ['column', 'column-reverse'];
var directions = ['ltr', 'rtl'];
var wraps = ['wrap', 'wrap-reverse'];
var linePacks = ['start', 'end', 'center', 'justify', 'distribute', 'stretch'];

writingModes.forEach(function(writingMode) {
    flexDirections.forEach(function(flexDirection) {
        directions.forEach(function(direction) {
            wraps.forEach(function(wrap) {
                linePacks.forEach(function(linePack) {
                    var flexboxClassName = writingMode + ' ' + direction + ' ' + flexDirection + ' ' + wrap + ' line-pack-' + linePack;
                    var title = document.createElement('div');
                    title.className = 'title';
                    title.innerHTML = flexboxClassName;
                    document.body.appendChild(title);

                    var mainAxis = 'height';
                    var crossAxis = 'width';

                    var flexbox = document.createElement('div');
                    flexbox.className = 'flexbox ' + flexboxClassName;
                    flexbox.setAttribute("data-expected-width", "600");
                    flexbox.setAttribute("data-expected-height", "20");

                    var testExpectations = expectations[writingMode][flexDirection][direction][wrap][linePack];
                    addChild(flexbox, 20, linePack, testExpectations[0]);
                    addChild(flexbox, 5, linePack, testExpectations[1]);

                    document.body.appendChild(flexbox);
                })
            })
        })
    })
})
</script>
</body>
</html>
